requirements.txt
appnope==0.1.3
asgiref==3.8.1
astroid==2.15.5
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
boto3==1.26.137
botocore==1.29.137
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.1.4
cryptography==40.0.2
debugpy==1.6.7.post1
decorator==5.1.1
dill==0.3.6
Django==5.0.7
django-cors-headers==4.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
executing==1.2.0
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
isort==5.12.0
jedi==0.19.0
jmespath==1.0.1
jupyter_client==8.3.0
jupyter_core==5.3.1
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
nest-asyncio==1.5.7
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg==3.2.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pykerberos==1.2.4
pylint==2.17.4
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-snappy==0.6.1
pytz==2023.3
pyzmq==25.1.1
requests==2.30.0
s3transfer==0.6.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.2
tomlkit==0.11.8
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.12.2
urllib3==1.26.15
wcwidth==0.2.6
wrapt==1.15.0
zstandard==0.21.0
